pyproject.toml
[project]
name = "shift-intensive-modeling"
version = "1.0.0"
description = "Detection project"
authors = [
    {name = "Alex Nerf",email = "a.varnakin@g.nsu.ru"}
]
license = {text = "No"}
readme = "README.md"
requires-python = ">=3.10.16,<3.14.0"
dependencies = [
    "numpy (>=1.26)",
    "hydra-colorlog (==1.2.0)",
    "hydra-core (==1.3.2)",
    "hydra-optuna-sweeper (==1.2.0)",
    "imageio (==2.37.0)",
    "ipykernel (==6.30.1)",
    "ipywidgets (==8.1.7)",
    "isoduration (==20.11.0)",
    "isort (==6.0.1)",
    "lightning (==2.5.5)",
    "lightning-utilities (==0.15.2)",
    "matplotlib (==3.10.6)",
    "matplotlib-inline (==0.1.7)",
    "mypy-extensions (==1.1.0)",
    "notebook (==7.4.7)",
    "notebook-shim (==0.2.4)",
    "nptyping (==2.5.0)",
    "omegaconf (==2.3.0)",
    "optuna (==2.10.1)",
    "orjson (==3.11.3)",
    "outcome (==1.3.0.post0)",
    "overrides (==7.7.0)",
    "packaging (==25.0)",
    "pandas (==2.3.3)",
    "pathlib (==1.0.1)",
    "pathspec (==0.12.1)",
    "pillow (==10.3.0)",
    "pkginfo (==1.12.1.2)",
    "platformdirs (==4.4.0)",
    "plotly (==6.0.0)",
    "poetry (==2.2.1)",
    "poetry-core (==2.2.1)",
    "pre-commit (==4.1.0)",
    "prometheus-client (==0.23.1)",
    "prompt-toolkit (==3.0.52)",
    "propcache (==0.4.1)",
    "pydantic (==2.12.0)",
    "pydantic-settings (==2.11.0)",
    "pydantic-core (==2.41.1)",
    "pyproject-hooks (==1.2.0)",
    "pytest (==8.3.5)",
    "python-doctr (==0.10.0)",
    "python-dotenv (==1.1.1)",
    "python-json-logger (==3.3.0)",
    "python-levenshtein (==0.27.1)",
    "pytorch-lightning (==2.5.5)",
    "pytz (==2025.2)",
    "pyviz-comms (==3.0.6)",
    "pyyaml (==6.0.3)",
    "pyzmq (==27.1.0)",
    "pyzstd (==0.17.0)",
    "rapidfuzz (==3.14.1)",
    "referencing (==0.36.2)",
    "requests (==2.32.5)",
    "requests-toolbelt (==1.0.0)",
    "rfc3339-validator (==0.1.4)",
    "rfc3986-validator (==0.1.1)",
    "rfc3987-syntax (==1.1.0)",
    "rich (==13.9.4)",
    "rich-argparse (==1.7.1)",
    "rootutils (==1.0.7)",
    "rpds-py (==0.27.1)",
    "ruamel-yaml (==0.18.15)",
    "ruamel-yaml-clib (==0.2.14)",
    "s3transfer (==0.14.0)",
    "safetensors (==0.6.2)",
    "scikit-image (==0.23.2)",
    "scipy (==1.15.3)",
    "scmrepo (==3.5.2)",
    "selenium (==4.24.0)",
    "semver (==3.0.4)",
    "send2trash (==1.8.3)",
    "shapely (==2.1.2)",
    "shellingham (==1.5.4)",
    "shortuuid (==1.0.13)",
    "shtab (==1.7.2)",
    "simsimd (==6.5.3)",
    "six (==1.17.0)",
    "smmap (==5.0.2)",
    "sniffio (==1.3.1)",
    "sortedcontainers (==2.4.0)",
    "soupsieve (==2.8)",
    "sqlalchemy (==2.0.43)",
    "sqltrie (==0.11.2)",
    "stack-data (==0.6.3)",
    "starlette (==0.46.2)",
    "stevedore (==5.5.0)",
    "stringzilla (==4.2.0)",
    "sympy (==1.13.1)",
    "tabulate (==0.9.0)",
    "tensorboard (==2.20.0)",
    "tensorboard-data-server (==0.7.2)",
    "timm (==1.0.20)",
    "tomli (==2.2.1)",
    "tomlkit (==0.13.3)",
    "tqdm (==4.67.1)",
    "typer (==0.19.2)",
    "types-python-dateutil (==2.9.0.20250822)",
    "typing-inspection (==0.4.2)",
    "typing-extensions (==4.15.0)",
    "ultralytics (==8.3.150)",
    "ultralytics-thop (==2.0.17)",
    "vine (==5.1.0)",
    "virtualenv (==20.34.0)",
    "torch (==2.6.0)",
    "torchvision (==0.21.0)",
    "torchmetrics (==1.8.2)",
    "opencv-python (==4.11.0.86)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pytorch-wheels/cu118/"
priority = "primary"

[tool.poetry.dependencies]
torch = {source = "aliyun"}
torchvision = {source = "aliyun"}


